+++ organisaatio-service/src/main/resources/META-INF/spring/bundle-context.xml
@@ -46,9 +46,9 @@
 
     <!-- Use following for testing purposes -->
     <!-- 
-    <bean id="ytjService" class="fi.vm.sade.rajapinnat.ytj.mock.YTJServiceMock" />
+    <bean id="ytjService" class="fi.vm.sade.organisaatio.ytj.mock.YTJServiceMock" />
     -->
-    <bean id="ytjService" class="fi.vm.sade.rajapinnat.ytj.service.YTJServiceImpl">
+    <bean id="ytjService" class="fi.vm.sade.organisaatio.ytj.service.YTJServiceImpl">
         <property name="asiakastunnus" value="${rajapinnat.ytj.asiakastunnus:YTJ_ASIAKASTUNNUS_UNKNOWN}"/>
         <property name="salainenavain" value="${rajapinnat.ytj.avain:YTJ_SALAINEN_AVAIN_UNKNOWN}"/>
     </bean>
