+++ organisaatio-service/src/test/java/fi/vm/sade/organisaatio/business/impl/OrganisaatioYtjServiceImplTest.java
@@ -5,10 +5,10 @@ import fi.vm.sade.organisaatio.api.model.types.OrganisaatioTyyppi;
 import fi.vm.sade.organisaatio.business.OrganisaatioYtjService;
 import fi.vm.sade.organisaatio.dao.OrganisaatioDAO;
 import fi.vm.sade.organisaatio.model.*;
-import fi.vm.sade.rajapinnat.ytj.api.YTJDTOBuilder;
-import fi.vm.sade.rajapinnat.ytj.api.YTJKieli;
-import fi.vm.sade.rajapinnat.ytj.api.YTJService;
-import fi.vm.sade.rajapinnat.ytj.api.exception.YtjConnectionException;
+import fi.vm.sade.organisaatio.ytj.api.YTJDTOBuilder;
+import fi.vm.sade.organisaatio.ytj.api.YTJKieli;
+import fi.vm.sade.organisaatio.ytj.api.YTJService;
+import fi.vm.sade.organisaatio.ytj.api.exception.YtjConnectionException;
 import org.assertj.core.groups.Tuple;
 import org.junit.After;
 import org.junit.Assert;
