+++ organisaatio-service/src/test/java/fi/vm/sade/organisaatio/business/impl/OrganisaatioYtjServiceImplValidationTest.java
@@ -4,8 +4,8 @@ import fi.vm.sade.oid.service.OIDService;
 import fi.vm.sade.oid.service.simple.OIDServiceSimpleImpl;
 import fi.vm.sade.organisaatio.business.OrganisaatioBusinessService;
 import fi.vm.sade.organisaatio.model.*;
-import fi.vm.sade.rajapinnat.ytj.api.YTJDTO;
-import fi.vm.sade.rajapinnat.ytj.api.YTJOsoiteDTO;
+import fi.vm.sade.organisaatio.ytj.api.YTJDTO;
+import fi.vm.sade.organisaatio.ytj.api.YTJOsoiteDTO;
 import fi.ytj.YTunnusDTO;
 import org.junit.Assert;
 import org.junit.Before;
