+++ pom.xml
@@ -19,7 +19,6 @@
         <koodisto.api.version>9.0-SNAPSHOT</koodisto.api.version>
         <log.client.version>8.0</log.client.version>
         <oid.api.version>8.0</oid.api.version>
-        <rajapinnat.ytj.version>13.0-SNAPSHOT</rajapinnat.ytj.version>
         <tarjonta.api.version>9.7-SNAPSHOT</tarjonta.api.version>
         <auditlogger.version>8.3.0-SNAPSHOT</auditlogger.version>
         <viestinta.version>14.8-SNAPSHOT</viestinta.version>
