--- ytj-client/ytj-api/src/main/java/fi/vm/sade/rajapinnat/ytj/api/YTJKieli.java
@@ -1,13 +1,0 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package fi.vm.sade.rajapinnat.ytj.api;
-
-/**
- *
- * @author tuomaskatva
- */
-public enum YTJKieli {
-    FI, SV, EN;
-}
