--- ytj-client/ytj-api/src/main/java/fi/vm/sade/rajapinnat/ytj/api/exception/YtjExceptionType.java
@@ -1,13 +1,0 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package fi.vm.sade.rajapinnat.ytj.api.exception;
-
-/**
- *
- * @author Tuomas Katva
- */
-public enum YtjExceptionType {
-    HTTP, SOAP, OTHER
-}
