+++ ytj-client/ytj-webservice/src/main/resources/META-INF/spring/context/bundle-context.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+
+    <bean id="ytjServiceImplBean" class="fi.vm.sade.organisaatio.ytj.service.YTJServiceImpl">
+        <property name="asiakastunnus" value="${rajapinnat.ytj.asiakastunnus}"/>
+        <property name="salainenavain" value="${rajapinnat.ytj.avain}"/>
+    </bean>
+
+</beans>
