+++ ytj-client/ytj-webservice/src/test/resources/log4j.properties
